Move functions out of main and into struct impl or view
