+++ src/s3.rs
@@ -8,6 +8,7 @@ pub(crate) async fn get_object(object_id: &str) -> Result<Vec<u8>> {
     // Given a filename, seek the file from S3 and return it as a byte stream.
 
     let config = Config::from_env();
+    println!("Config: {:?}", config);
     let path = format!("{}-{}", config.app_name, config.deployment);
     let credentials = Credentials::new(
         Some(&config.s3_access_key),
@@ -21,8 +22,8 @@ pub(crate) async fn get_object(object_id: &str) -> Result<Vec<u8>> {
         region: config.s3_region,
         endpoint: config.s3_endpoint,
     };
-
     let filename = format!("{}/{}", path, object_id);
+    println!("Getting object: {}", filename);
     let bucket = Bucket::new(&config.s3_bucket_id, region, credentials)?;
 
     let response_data = bucket.get_object(filename).await?.bytes().to_vec();
